requirements.txt
# Telegram Bot
python-telegram-bot==20.7
python-telegram-bot[job-queue]==20.7

# Scheduling
APScheduler==3.10.4

# AI Integration
anthropic==0.34.2

# Configuration
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Markdown parsing
markdown==3.7
PyYAML==6.0.2

# Timezone handling
pytz==2024.2

# Utilities
requests==2.32.3

# HTTP Server (for Cloud Run)
Flask==3.0.0

# Testing (dev dependencies)
pytest==8.3.3
pytest-cov==5.0.0
pytest-asyncio==0.24.0
